requirements.txt

streamlit==1.31.1
beautifulsoup4>=4.12.2
langchain-experimental
bs4>=0.0.1
lxml>=4.9.3
requests>=2.31.0
python-dotenv>=1.0.0
langchain>=0.0.350
langchain-core>=0.1.18
langchain-community>=0.0.10
langchain-google-genai>=0.0.3
langchainhub>=0.1.13
chromadb>=0.4.17
rank-bm25>=0.2.2
sentence-transformers>=2.2.2
spacy>=3.7.2
networkx>=3.2.1
PyPDF2>=3.0.0
html2text>=2024.2.25
langchain-text-splitters>=0.0.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz